try to add a volume for the database

+++ src/requirements/wordpress/conf/auto_config.sh
@@ -4,14 +4,24 @@ sleep 10
 
 cd /var/www/html
 wp core download --allow-root
+# rm -rf /var/www/html/*
 
-# replace sed command with wp config set 
-sed -i '23 s/database_name_here/maria/' /var/www/html/wp-config-sample.php
-sed -i '26 s/username_here/aa/' /var/www/html/wp-config-sample.php
-sed -i '29 s/password_here/aa/' /var/www/html/wp-config-sample.php
-sed -i '32 s/localhost/db-service:3306/' /var/www/html/wp-config-sample.php
-sed -i '82 s/false/true/' /var/www/html/wp-config-sample.php
-mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+if [ ! -f "wp-config.php" ]; then
+    mv /var/www/html/wp-config-sample.php /var/www/html/wp-config.php
+fi
+
+# wp config set DB_NAME maria --allow-root --path='/var/www/html'
+# wp config set DB_USER aa --allow-root --path='/var/www/html'
+# wp config set DB_PASSWORD aa --allow-root --path='/var/www/html'
+# wp config set DB_HOST db-service:3306 --allow-root --path='/var/www/html'
+# wp config set WP_DEBUG true --allow-root --path='/var/www/html'
+
+sed -i '23 s/database_name_here/maria/' /var/www/html/wp-config.php
+sed -i '26 s/username_here/aa/' /var/www/html/wp-config.php
+sed -i '29 s/password_here/aa/' /var/www/html/wp-config.php
+sed -i '32 s/localhost/maria:3306/' /var/www/html/wp-config.php
+# sed -i '82 s/false/true/' /var/www/html/wp-config.php
+
 # Create directory for sock-file
 mkdir -p /run/php
 
@@ -25,4 +35,3 @@ wp core install --url="localhost" --title="My Wordpress Site" --admin_user="eken
 # fi
 
 /usr/sbin/php-fpm7.4 -F
-# /usr/sbin/php-fpm7.4
